third_party/WebKit/LayoutTests/http/tests/cache/smil-usecounter-in-cached-image-expected.txt

CONSOLE WARNING: SVG's SMIL animations (<animate>, <set>, etc.) are deprecated and will be removed. Please use CSS animations or Web animations instead.
CONSOLE WARNING: SVG's SMIL animations (<animate>, <set>, etc.) are deprecated and will be removed. Please use CSS animations or Web animations instead.
PASS window.internals.isUseCounted(document, SVGSMILAnimationInImageRegardlessOfCache) is false
PASS window.internals.isUseCounted(document, SVGSMILAnimationInImageRegardlessOfCache) is true
PASS successfullyParsed is true

TEST COMPLETE
